pyproject.toml
# PEP 518 / PEP 621
# https://packaging.python.org/en/latest/specifications/pyproject-toml/

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ILThermoML"
requires-python = ">=3.12,<3.13"
dynamic = ["version"]
dependencies = [
  "ilthermopy>=1.1.0",
  "joblib>=1.4.2",
  "pandas>=2.2.3",
  "python-dotenv>=1.0.1",
  "tqdm>=4.67.1",
  "types-tqdm>=4.67.0.20250301",
]

[dependency-groups]
local = ["ipykernel>=6.29.5", "pre-commit>=4.1.0"]
mypy = ["mypy>=1.14.1"]
ruff = ["ruff>=0.9.2"]
test = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "pytest-env>=1.1.5",
  "pytest-mock>=3.14.0",
  "pytest-randomly>=3.16.0",
  "semver>=3.0.2",
]

# uv
# https://docs.astral.sh/uv/reference/settings/

[tool.uv]
package = true
required-version = ">=0.5.21"

# setuptools
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html

[tool.setuptools.dynamic]
version = { attr = "ilthermoml.__version__" }

# Ruff
# https://docs.astral.sh/ruff/settings/
# https://docs.astral.sh/ruff/rules/

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
ignore = ["COM812", "D1", "D205"]
select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F403"]
"**/test*.py" = ["ANN", "S101"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["ilthermoml"]

# mypy
# https://mypy.readthedocs.io/en/stable/config_file.html

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["ilthermopy.*", "joblib", "pandas", "pytest", "pytest_mock", "semver"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_decorators = false

# pytest & plugins
# https://docs.pytest.org/en/stable/reference/reference.html#ini-options-ref
# https://pytest-cov.readthedocs.io/en/latest/config.html#reference
# https://github.com/pytest-dev/pytest-env

[tool.pytest.ini_options]
env = ["TQDM_DISABLE=true"]
addopts = [
  "--cov=src/",
  "--cov-branch",
  "--cov-report=term-missing:skip-covered",
]
python_files = ["test*.py"]
testpaths = ["tests/"]

[tool.coverage.report]
exclude_also = ["@(abc\\.)?abstractmethod", "if TYPE_CHECKING:"]
